requirements.txt
matplotlib==3.10.8
numpy==2.4.1
Pillow==12.1.0
schemdraw==0.22
setuptools==65.5.0
wavedrom==2.0.3.post3
